# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/recipients": {
      "get": {
        "operationId": "RecipientsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all recipients.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipientResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all recipients",
        "tags": [
          "recipients"
        ]
      },
      "post": {
        "operationId": "RecipientsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data to create a new recipient.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The recipient has been created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipientCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — the slug or the tax ID is already taken."
          }
        },
        "summary": "Create a new recipient",
        "tags": [
          "recipients"
        ]
      }
    },
    "/recipients/{id}": {
      "get": {
        "operationId": "RecipientsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Recipient ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipient found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipientResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recipient not found."
          }
        },
        "summary": "Get a single recipient",
        "tags": [
          "recipients"
        ]
      },
      "patch": {
        "operationId": "RecipientsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the recipient to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New data for the recipient.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recipient updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipientCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — the slug or the tax ID is already taken."
          },
          "404": {
            "description": "Recipient not found."
          }
        },
        "summary": "Update a recipient",
        "tags": [
          "recipients"
        ]
      },
      "delete": {
        "operationId": "RecipientsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Recipient ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipient deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recipient not found."
          }
        },
        "summary": "Delete a recipient by its ID",
        "tags": [
          "recipients"
        ]
      }
    },
...
    "/locations": {
      "get": {
        "operationId": "LocationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all locations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all locations",
        "tags": [
          "locations"
        ]
      },
      "post": {
        "operationId": "LocationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data to create a new location.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The location has been created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — owner is not exactly one of retailer/recipient, or the owner already has a primary location."
          }
        },
        "summary": "Create a new location",
        "tags": [
          "locations"
        ]
      }
    },
    "/locations/{id}": {
      "get": {
        "operationId": "LocationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Location ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Location not found."
          }
        },
        "summary": "Get a single location",
        "tags": [
          "locations"
        ]
      },
      "patch": {
        "operationId": "LocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the location to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New data for the location.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — owner is not exactly one of retailer/recipient, or the owner already has a primary location."
          },
          "404": {
            "description": "Location not found."
          }
        },
        "summary": "Update a location",
        "tags": [
          "locations"
        ]
      },
      "delete": {
        "operationId": "LocationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Location ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Location not found."
          }
        },
        "summary": "Delete a location by its ID",
        "tags": [
          "locations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Spira API Documentation",
    "description": "Spira API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FoodCategory": {
        "type": "string",
        "enum": [
          "PRODUCE",
          "BAKERY",
          "DAIRY",
          "MEAT",
          "PREPARED",
          "DRY_GOODS",
          "FROZEN"
        ],
        "description": "Food categories typically donated."
      },
...
      "ProfileStatus": {
        "type": "string",
        "enum": [
          "PENDING_VERIFICATION",
          "ACTIVE",
          "SUSPENDED",
          "INACTIVE"
        ],
        "description": "Lifecycle status."
      },
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable message describing the outcome.",
            "example": "Request completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "RecipientType": {
        "type": "string",
        "enum": [
          "NGO",
          "FOOD_BANK",
          "SOUP_KITCHEN",
          "SHELTER",
          "COMMUNITY_FRIDGE",
          "CHURCH",
          "SCHOOL",
          "CERTIFIED_INDIVIDUAL"
        ],
        "description": "What kind of receiver this is."
      },
      "DietaryRestriction": {
        "type": "string",
        "enum": [
          "HALAL",
          "KOSHER",
          "NO_PORK",
          "NO_ALCOHOL",
          "VEGETARIAN"
        ],
        "description": "Dietary rules their beneficiaries follow."
      },
      "BackgroundCheckStatus": {
        "type": "string",
        "enum": [
          "NOT_REQUIRED",
          "PENDING",
          "PASSED",
          "FAILED"
        ],
        "description": "Background-check outcome."
      },
      "RecipientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique recipient ID.",
            "format": "uuid"
          },
          "type": {
            "description": "What kind of receiver this is.",
            "example": "FOOD_BANK",
            "allOf": [
              {
                "$ref": "#/components/schemas/RecipientType"
              }
            ]
          },
          "legalName": {
            "type": "object",
            "description": "Registered organization name.",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "description": "Public name of the organization, or the person's name.",
            "example": "Banco de Alimentos Paraguay"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier."
          },
          "taxId": {
            "type": "object",
            "description": "Government tax identifier.",
            "nullable": true
          },
          "nonprofitRegistrationNumber": {
            "type": "object",
            "description": "Nonprofit registration number.",
            "nullable": true
          },
          "nationalId": {
            "type": "object",
            "description": "National ID document. Individuals only.",
            "nullable": true
          },
          "mission": {
            "type": "object",
            "description": "Mission statement.",
            "nullable": true
          },
          "website": {
            "type": "object",
            "description": "Public website.",
            "nullable": true
          },
          "logoUrl": {
            "type": "object",
            "description": "Absolute URL of the logo.",
            "nullable": true
          },
          "serviceRadiusKm": {
            "type": "number",
            "description": "How far they are willing to travel, in kilometres.",
            "nullable": true
          },
          "hasVehicle": {
            "type": "boolean",
            "description": "Whether they have their own vehicle."
          },
          "transportCapacityKg": {
            "type": "number",
            "description": "How much they can carry in one trip, in kilograms.",
            "nullable": true
          },
          "hasRefrigeratedTransport": {
            "type": "boolean",
            "description": "Whether their transport is refrigerated."
          },
          "peopleServedPerWeek": {
            "type": "number",
            "description": "People fed per week.",
            "nullable": true
          },
          "maxDailyIntakeKg": {
            "type": "number",
            "description": "Most they can take in a single day, in kilograms.",
            "nullable": true
          },
          "acceptedFoodCategories": {
            "type": "array",
            "description": "Categories they will take.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            }
          },
          "excludedFoodCategories": {
            "type": "array",
            "description": "Categories they explicitly refuse.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            }
          },
          "dietaryRestrictions": {
            "type": "array",
            "description": "Dietary rules their beneficiaries follow.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/DietaryRestriction"
            }
          },
          "acceptsNearExpiry": {
            "type": "boolean",
            "description": "Whether they accept near-expiry food."
          },
          "acceptsPreparedFood": {
            "type": "boolean",
            "description": "Whether they accept prepared food."
          },
          "acceptsFrozen": {
            "type": "boolean",
            "description": "Whether they accept frozen goods."
          },
          "status": {
            "description": "Lifecycle status.",
            "example": "PENDING_VERIFICATION",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileStatus"
              }
            ]
          },
          "isVerified": {
            "type": "boolean",
            "description": "Whether an admin has verified this recipient."
          },
          "verifiedAt": {
            "type": "string",
            "description": "When the recipient was verified (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "verifiedById": {
            "type": "object",
            "description": "ID of the admin who verified this recipient.",
            "format": "uuid",
            "nullable": true
          },
          "foodHandlingCertificationNumber": {
            "type": "object",
            "description": "Food-handling certification number.",
            "nullable": true
          },
          "certificationExpiresAt": {
            "type": "string",
            "description": "Certification expiry date.",
            "format": "date",
            "nullable": true
          },
          "backgroundCheckStatus": {
            "description": "Background-check outcome.",
            "example": "NOT_REQUIRED",
            "allOf": [
              {
                "$ref": "#/components/schemas/BackgroundCheckStatus"
              }
            ]
          },
          "insurancePolicyNumber": {
            "type": "object",
            "description": "Liability insurance policy number.",
            "nullable": true
          },
          "termsAcceptedAt": {
            "type": "string",
            "description": "When the platform terms were accepted (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "termsVersion": {
            "type": "object",
            "description": "Version of the terms accepted.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When the recipient was created (ISO 8601).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the recipient was last modified (ISO 8601).",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "displayName",
          "slug",
          "hasVehicle",
          "hasRefrigeratedTransport",
          "acceptsNearExpiry",
          "acceptsPreparedFood",
          "acceptsFrozen",
          "status",
          "isVerified",
          "backgroundCheckStatus",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateRecipientDto": {
        "type": "object",
        "properties": {
          "type": {
            "description": "What kind of receiver this is.",
            "example": "FOOD_BANK",
            "allOf": [
              {
                "$ref": "#/components/schemas/RecipientType"
              }
            ]
          },
          "legalName": {
            "type": "string",
            "description": "Registered organization name. Leave unset for a certified individual.",
            "example": "Fundación Banco de Alimentos Paraguay",
            "maxLength": 200
          },
          "displayName": {
            "type": "string",
            "description": "Public name of the organization, or the person's name.",
            "example": "Banco de Alimentos Paraguay",
            "minLength": 2,
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier. Must be unique.",
            "example": "banco-de-alimentos-py",
            "maxLength": 120
          },
          "taxId": {
            "type": "string",
            "description": "Government tax identifier. Organizations only, and unique.",
            "example": "80098765-4",
            "maxLength": 40
          },
          "nonprofitRegistrationNumber": {
            "type": "string",
            "description": "Registration number proving nonprofit status.",
            "maxLength": 80
          },
          "nationalId": {
            "type": "string",
            "description": "National ID document. Individuals only.",
            "maxLength": 40
          },
          "mission": {
            "type": "string",
            "description": "Free-text description of who they serve and how."
          },
          "website": {
            "type": "string",
            "description": "Public website.",
            "maxLength": 255
          },
          "logoUrl": {
            "type": "string",
            "description": "Absolute URL of the logo.",
            "maxLength": 255
          },
          "serviceRadiusKm": {
            "type": "number",
            "description": "How far they are willing to travel, in kilometres.",
            "example": 15,
            "minimum": 0,
            "maximum": 5000
          },
          "hasVehicle": {
            "type": "boolean",
            "description": "Whether they have their own vehicle for collections.",
            "default": false
          },
          "transportCapacityKg": {
            "type": "number",
            "description": "How much they can carry in one trip, in kilograms.",
            "example": 750.5,
            "minimum": 0,
            "maximum": 999999.99
          },
          "hasRefrigeratedTransport": {
            "type": "boolean",
            "description": "Whether their transport is refrigerated.",
            "default": false
          },
          "peopleServedPerWeek": {
            "type": "number",
            "description": "People fed per week.",
            "example": 1200,
            "minimum": 0
          },
          "maxDailyIntakeKg": {
            "type": "number",
            "description": "Most they can take in a single day, in kilograms.",
            "example": 2000,
            "minimum": 0,
            "maximum": 999999.99
          },
          "acceptedFoodCategories": {
            "type": "array",
            "description": "Categories they will take.",
            "example": [
              "PRODUCE",
              "DRY_GOODS"
            ],
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            }
          },
          "excludedFoodCategories": {
            "type": "array",
            "description": "Categories they explicitly refuse.",
            "example": [
              "MEAT"
            ],
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            }
          },
          "dietaryRestrictions": {
            "type": "array",
            "description": "Dietary rules their beneficiaries follow.",
            "example": [
              "NO_PORK"
            ],
            "items": {
              "$ref": "#/components/schemas/DietaryRestriction"
            }
          },
          "acceptsNearExpiry": {
            "type": "boolean",
            "description": "Whether they accept food close to its expiry date.",
            "default": false
          },
          "acceptsPreparedFood": {
            "type": "boolean",
            "description": "Whether they accept already-prepared food.",
            "default": false
          },
          "acceptsFrozen": {
            "type": "boolean",
            "description": "Whether they accept frozen goods.",
            "default": false
          },
          "foodHandlingCertificationNumber": {
            "type": "string",
            "description": "Food-handling certification number.",
            "maxLength": 80
          },
          "certificationExpiresAt": {
            "type": "string",
            "description": "Certification expiry date (calendar date, no time).",
            "format": "date",
            "example": "2027-06-30"
          },
          "insurancePolicyNumber": {
            "type": "string",
            "description": "Liability insurance policy number.",
            "maxLength": 80
          },
          "termsAcceptedAt": {
            "type": "string",
            "description": "When the recipient accepted the platform terms.",
            "format": "date-time",
            "example": "2026-09-16T14:32:05.123Z"
          },
          "termsVersion": {
            "type": "string",
            "description": "Version of the terms that was accepted.",
            "example": "2026-01",
            "maxLength": 20
          }
        },
        "required": [
          "type",
          "displayName",
          "slug"
        ]
      },
...
    }
  }
}